Account creation form, CurrencyUtils, Entity modification
New form was added into overview page of customers. It provides form to create new account with specified currency. For this purpuse CurrencyUtils was created. Utils are based on javamoney library and works with BigDecimal datatype so it was necessary to modify Account and Transaction Entity to support this datatype.